+++ test/facebook.js
@@ -2,12 +2,19 @@ import { createFacebookVideo } from './lib/util'
 import assert from 'assert'
 
 describe('Facebook Video', () => {
-  it('should provide the facebook poster image', async () => {
-    const { query } = await createFacebookVideo('10156049485672318')
-    assert.strictEqual(
-      query('img').getAttribute('src'),
-      'video/facebook/10156049485672318-poster-image'
-    )
+  it('should provide no poster image', async () => {
+    const { query } = await createFacebookVideo('10156049485672318', {
+      'poster-image': '',
+    })
+    assert.ok(!query('img'))
+  })
+
+  it('should provide the given poster image', async () => {
+    const posterImage = 'http://www.test.de/facebook'
+    const { query } = await createFacebookVideo('10156049485672318', {
+      'poster-image': posterImage,
+    })
+    assert.strictEqual(query('img').getAttribute('src'), posterImage)
   })
 
   it('should load the facebook player after click', async () => {